# Fix: sum to double not int64 for avg(int) without group by