## What type of PR is this：
- [x] bug
- [ ] feature
- [ ] enhancement
- [ ] others

## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #5201.

## Problem Summary：
<!-- (Please describe the changes you have made. In which scenarios will this bug be triggered and what measures have you taken to fix the bug?) -->
As for `SELECT AVG(bigint)` without group by, `AvgAggregateFunction::update_batch_single_state` speedd up the local sum operations as follows:
1. local_sum_for_arithmetic += value in loop,   `local_sum_for_arithmetic` and `value` is `int64_t`.
2. this->data(state).sum += local_sum_for_arithmetic, `sum` is `double`.

However, the value range of `int64` is smaller than `double`, so for this case, `double` isn't overflow, but `int64` will be overflow.

I tried to use `local_sum_for_arithmetic` firstly, and turn back to use `double sum` when it overflows. However, checking overflow in loop makes the code not able to vectorized. See [Compiler Explorer](https://godbolt.org/z/q7rG5hTWb) for detail.

## Test
### Environment
- 3 BE * 16vCPUs.
- ssb_100g
- Query: `SELECT AVG(lo_orderkey) FROM lineorder`
- Enable pipeline, and `pipeline_dop=1`.

### Result
Aggregate performance decreases about 3 times than before.
- Before PR. Latency: 140ms, Aggregate costs: 96ms.
- After PR. Latency: 300ms, Aggregate costs: 271ms.