@ZiheLiu 2022-04-21
run starrocks_be_unittest