@ZiheLiu 2022-04-21
run starrocks_clang-format